Move violation service test to a separate file
